Move supporting functions and types to files linebuffer.py and diagnostics.py
